refactor: move page to grouped routes
